+++ src/App.scss
@@ -1,0 +1,9 @@
+#root {
+  width: 100vw;
+  margin: 0;
+}
+
+body {
+  box-sizing: border-box;
+  margin: 0;
+}
