+++ src/components/CommunityCards/CommunityCards.jsx
@@ -7,29 +7,29 @@ import SoloCommCard from '../SoloCommCard/SoloCommCard'
 import './CommunityCards.scss'
 
 function CommunityCards() {
-    const url = import.meta.env.VITE_API_URL;
+    // const url = import.meta.env.VITE_API_URL;
 
-    // STATE VARIABLES FOR API CALL
-    const [communities, setCommunities] = useState([]);
-    const [error, setError] = useState(null);
+    // // STATE VARIABLES FOR API CALL
+    // const [communities, setCommunities] = useState([]);
+    // const [error, setError] = useState(null);
 
-    useEffect(() => {
-        // AXIOS GET
-        const getCommunities = async () => {
-            try {
-                const response = await axios.get(`${url}/users`);
-                console.log(response.data);
-                // need communities for user
+    // useEffect(() => {
+    //     // AXIOS GET
+    //     const getCommunities = async () => {
+    //         try {
+    //             const response = await axios.get(`${url}/users`);
+    //             console.log(response.data);
+    //             // need communities for user
 
-                setCommunities(response.data);
-            } catch (error) {
-                console.error(error);
-                setError(error);
-            }
-        };
-        // LOAD API DATA ONTO SCREEN
-        getCommunities();
-    }, []); // [] = runs once + right away!
+    //             setCommunities(response.data);
+    //         } catch (error) {
+    //             console.error(error);
+    //             setError(error);
+    //         }
+    //     };
+    //     // LOAD API DATA ONTO SCREEN
+    //     getCommunities();
+    // }, []); // [] = runs once + right away!
 
 
     return (
